Adds refresh for news feed
